Enabled UserContoller class test cases. Moved the logic of validating password from user service to user contoller
